test: 주소 관련 로직 테스트 코드 추가

+++ api-module/src/main/java/com/project/address/service/AddressServiceImpl.java
@@ -3,6 +3,7 @@ package com.project.address.service;
 import com.project.address.dto.AddressRequestDto;
 import com.project.address.dto.AddressResponseDto;
 import com.project.common.ApiResponse;
+import com.project.common.exception.AddressException;
 import com.project.domain.address.Addresses;
 import com.project.domain.address.repository.AddressRepository;
 import lombok.RequiredArgsConstructor;
@@ -24,13 +25,15 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public ApiResponse getAddress(Long id) {
-        return ApiResponse.success(addressRepository.findById(id));
+        Addresses address = addressRepository.findById(id).orElseThrow(() -> new AddressException("Address not found with id: " + id));
+
+        return ApiResponse.success(addressConverter.convertAddressEntityToResponse(address));
     }
 
     @Override
     public ApiResponse updateAddress(AddressRequestDto request) {
         Addresses address = addressRepository.findById(request.getAddressId())
-                .orElseThrow(() -> new RuntimeException("Address not found with id " + request.getAddressId()));
+                .orElseThrow(() -> new AddressException("Address not found with id " + request.getAddressId()));
 
         Addresses updateAddress = addressConverter.updateAddressFromRequest(address, request);
 
@@ -41,6 +44,9 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public void deleteAddress(Long id) {
+        if (!addressRepository.existsById(id)) {
+            throw new AddressException("Address not found with id: " + id);
+        }
         addressRepository.deleteById(id);
     }
 }
